+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed broken line charts
 - fixed a problem with changing the settings of single value component
 - zoom reset btn for all multilines
+- zoom reset on fullscreen
 - reduced reset zoom time